Refactor animation to seprate component.
